feat: moved helper functons to the utils
